backend/requirements.txt
django
djangorestframework
firebase-admin
psycopg2-binary
python-dotenv
django-cors-headers
gunicorn
whitenoise
dj-database-url
numpy
pandas
scikit-learn
google-generativeai
deep-translator
groq
